refactor: AI 요약 유틸 통합 및 섹션 컴포넌트 경로 정리
